Add app icon, added a trigger URL and changed some entities so that we can create a trigger URL from the stored data

+++ CountdownViewController.m
@@ -51,7 +51,7 @@ const NSInteger TOTAL_COUNTDOWN = 5;
 -(void) updateCountdown {
     
     self.countdownRemaining--;
-    [self.lblCountdown setText:[NSString stringWithFormat:@"%ld", self.countdownRemaining]];
+    [self.lblCountdown setText:[NSString stringWithFormat:@"%ld", (long)self.countdownRemaining]];
     
     if (self.countdownRemaining <= 0)
     {
